+++ Tests/Fenrir.Application.Game.Tests/Social/TradeCommitPlannerTests.cs
@@ -15,25 +15,25 @@ public class TradeCommitPlannerTests
     public void BuildFinalContainers_RemovesOwnOfferedItem_AndAddsReceivedItemToFreeSlot()
     {
         var page0 = ImmutableDictionary<byte, ItemStack>.Empty
-            .Add((byte)0, Stack(100)) // offered away
-            .Add((byte)1, Stack(200)); // kept
+            .Add(0, Stack(100)) // offered away
+            .Add(1, Stack(200)); // kept
 
         var page1 = ImmutableDictionary<byte, ItemStack>.Empty;
 
         var ownOffered = new (byte, byte, ItemStack)?[]
         {
-            (ContainerMatrix.InventoryPage0, (byte)0, Stack(100))
+            (ContainerMatrix.InventoryPage0, 0, Stack(100))
         };
         var received = new (byte, byte, ItemStack)?[]
         {
-            (ContainerMatrix.InventoryPage0, (byte)5, Stack(999)) // source container/slot irrelevant on the receiving side
+            (ContainerMatrix.InventoryPage0, 5, Stack(999)) // source container/slot irrelevant on the receiving side
         };
 
         var plan = TradeCommitPlanner.BuildFinalContainers(page0, page1, ownOffered, received);
 
         Assert.False(plan.Overflowed);
-        Assert.Equal(200, plan.Page0[(byte)1].ItemId); // untouched
-        Assert.Equal(999, plan.Page0[(byte)0].ItemId); // slot 0 vacated by the offered item, then reused for the received one
+        Assert.Equal(200, plan.Page0[1].ItemId); // untouched
+        Assert.Equal(999, plan.Page0[0].ItemId); // slot 0 vacated by the offered item, then reused for the received one
     }
 
     [Fact]
@@ -45,12 +45,12 @@ public class TradeCommitPlannerTests
         var page0 = builder.ToImmutable();
         var page1 = ImmutableDictionary<byte, ItemStack>.Empty;
 
-        var received = new (byte, byte, ItemStack)?[] { (ContainerMatrix.InventoryPage0, (byte)0, Stack(999)) };
+        var received = new (byte, byte, ItemStack)?[] { (ContainerMatrix.InventoryPage0, 0, Stack(999)) };
 
         var plan = TradeCommitPlanner.BuildFinalContainers(page0, page1, [], received);
 
         Assert.False(plan.Overflowed);
-        Assert.Equal(999, plan.Page1[(byte)0].ItemId);
+        Assert.Equal(999, plan.Page1[0].ItemId);
     }
 
     [Fact]
@@ -64,7 +64,7 @@ public class TradeCommitPlannerTests
             builder1.Add((byte)i, Stack(1));
         }
 
-        var received = new (byte, byte, ItemStack)?[] { (ContainerMatrix.InventoryPage0, (byte)0, Stack(999)) };
+        var received = new (byte, byte, ItemStack)?[] { (ContainerMatrix.InventoryPage0, 0, Stack(999)) };
 
         var plan = TradeCommitPlanner.BuildFinalContainers(builder0.ToImmutable(), builder1.ToImmutable(), [],
             received);
